requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.12.3
pydantic-settings==2.1.0
pydantic[email]==2.12.3
bcrypt==4.1.2
python-jose==3.3.0
cryptography==42.0.2
email-validator==2.1.0
